Fix to [](c) and file move
